moved CLI and init functions in Init.ts
